Extracted drag/drop list to its own component
