+++ titan-web-client/src/components/members/form/AddMembersForm.js
@@ -1,31 +1,22 @@
 import React, { useEffect, useState } from 'react';
 import Button from '@material-ui/core/Button';
-import TextField from '@material-ui/core/TextField';
-import InputAdornment from '@material-ui/core/InputAdornment';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { useThrottle } from 'titan/hooks';
 import {
   AddUserToOrganizationRequest,
-  ListUsersRequest,
-  makeTitanApiRequest,
-  useTitanApiClient
+  makeTitanApiRequest
 } from 'titan/http/ApiClient';
-import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import Column from 'titan/components/Grid/Column';
 import Avatar from '@material-ui/core/Avatar';
 import Row from 'titan/components/Grid/Row';
 import { Typography } from '@material-ui/core';
 import { useSnackbar } from 'notistack';
+import { SearchMembersForm } from 'titan/components/members/form/SearchMembersForm';
 
 /**
  * @param {{orgId, orgCoc, users}} props
  */
 export function AddMembersForm (props) {
   const [addedUsers, setAddedUsers] = useState([]);
-  const [username, setUsername] = useThrottle('', 325);
-  const fetchUsers = useTitanApiClient(
-    ListUsersRequest, { username, limit: 5 });
   const snackbar = useSnackbar();
 
   useEffect(() => {
@@ -50,45 +41,31 @@ export function AddMembersForm (props) {
   }
 
   return (
-    <div>
-      <TextField
-        fullWidth
-        InputProps={{
-          startAdornment: (
-            <InputAdornment position="start">
-              <FontAwesomeIcon icon="search" />
-            </InputAdornment>
-          )
-        }}
-        placeholder="Search by username"
-        onChange={e => setUsername(e.target.value)}
-      />
-      <List dense style={{ height: 285 }}>
-        {fetchUsers.data && fetchUsers.data.map((user, index) => (
-          <ListItem key={index} button>
-            <Row alignItems="center" alignContent="center">
-              <Column>
-                <Avatar
-                  style={{ width: 20, height: 20 }}
-                  src={user.wcf.avatar_url}
-                />
-              </Column>
-              <Column>{user.username}</Column>
-              <Column grow={1}>
-                <Typography align="right">
-                  {addedUsers.includes(user.id) ? (
-                    <Button disabled color="primary">Added</Button>
-                  ) : (
-                    <Button
-                      color="primary"
-                      onClick={() => addMember(user.id)}>Add</Button>
-                  )}
-                </Typography>
-              </Column>
-            </Row>
-          </ListItem>
-        ))}
-      </List>
-    </div>
+    <SearchMembersForm itemRenderer={(user, index) => (
+      <ListItem button component="div" key={index}>
+        <Row alignItems="center" alignContent="center">
+          <Column>
+            <Avatar
+              component="div"
+              style={{ width: 20, height: 20 }}
+              src={user.wcf.avatar_url}
+            />
+          </Column>
+          <Column>{user.username}</Column>
+          <Column grow={1}>
+            <Typography align="right">
+              {addedUsers.includes(user.id) ? (
+                <Button disabled color="primary">Added</Button>
+              ) : (
+                <Button
+                  color="primary"
+                  onClick={() => addMember(user.id)}>Add</Button>
+              )}
+            </Typography>
+          </Column>
+        </Row>
+      </ListItem>
+    )}
+    />
   );
 }
